LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.7_Punctuators/S7.7_A2_T7-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '\u002B'
S7.7_A2_T7

PASS Expected parsing failure

TEST COMPLETE
